requirements.txt
openai>=0.27.0
python-pptx>=0.6.21
python-dotenv>=0.19.0 